requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
black==25.11.0
click==8.3.0
Faker==38.0.0
fastapi==0.121.2
flake8==7.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.11
py4j==0.10.9.9
pyarrow==22.0.0
pycodestyle==2.14.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pyspark==4.0.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
httpx==0.28.1
